+++ packages/network/src/context.ts
@@ -10,8 +10,8 @@ export interface SodiumContext {
 const defaultSodiumContext: SodiumContext = {
   factoryAddress: "0x5Eb4bcAEB78a7a765FA3F7285eE6B3E08CCB2c09",
   sodiumAuthAddress: "0x8987ECEa30337FD2ebC5A07Da0cc6EFFCc051BE6",
-  singletonAddress: "0x10672E5954092c08E20917d959A7735781d221C8",
-  fallbackHandlerAddress: "0x1aC8A9d1a4bD87bC2853358d7Bc9ABd8B49a8F7D",
+  singletonAddress: "0xC025d37f998508b8eb566e5648164c2698Bb9643",
+  fallbackHandlerAddress: "0x1fD9348984a887cE6bDb2f2931eda0EA1427C422",
   opValidatorAddress: "0x56e049F8bf238233B1eB93943fbE629B8B99B089",
   entryPointAddress: "0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789",
 };